% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsim.R
\name{breed_one_year_fonly}
\alias{breed_one_year_fonly}
\title{Breed 1 year for females only}
\usage{
breed_one_year_fonly(population, year_now, n_babies)
}
\arguments{
\item{population}{the population to breed}
}
\description{
Select mature inviduals breed them together to create new kids!
}
