% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsim.R
\name{breed_one_year_od}
\alias{breed_one_year_od}
\title{Breed one year with 2 sexes, overdispersion and sex ratios}
\usage{
breed_one_year_od(population, year_now, overdispersion, sex_fraction)
}
\arguments{
\item{population}{the population to breed}

\item{year_now}{the current year}

\item{overdispersion}{the vector with the overdispersion parameter for each sex}

\item{sex_fraction}{fraction of population that is male}
}
\description{
Breed one year with 2 sexes, overdispersion and sex ratios
}
