% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsim.R
\name{create_founding_pop}
\alias{create_founding_pop}
\title{Create a founding population}
\usage{
create_founding_pop(init_N, markers, maturity_fun, surv_fun, fecund_fun)
}
\arguments{
\item{init_N}{dataframe with the initial population details, must be Age, Sex then N}

\item{markers}{dataframe with information on markers must be geno, locus then freqs}

\item{maturity_fun}{The function that specifies at what age they become mature}

\item{surv_fun}{the survival function}

\item{fecund_fun}{the mean fecundity at age function}
}
\description{
Create a founding population
}
