% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsim.R
\name{deaths_breed_one_year}
\alias{deaths_breed_one_year}
\title{Deaths and then randomly breed 2 sexes}
\usage{
deaths_breed_one_year(population, year_now, n_babies)
}
\arguments{
\item{population}{the population to die and breed in}

\item{year_now}{the current year of the simulation}

\item{n_babies}{function describing the distribution how offspring are generated}
}
\description{
Deaths and then randomly breed 2 sexes
}
