% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simhelpers.R
\name{fecund_inv_fun}
\alias{fecund_inv_fun}
\title{Simple fecundity at age function}
\usage{
fecund_inv_fun(age, fpar)
}
\arguments{
\item{age}{the age of the individual}

\item{fpar}{the fecundity parameter}
}
\description{
Simple fecundity at age function
}
