% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsim.R
\name{generate_kids_od}
\alias{generate_kids_od}
\title{Generate kids randomly with overdispersion}
\usage{
generate_kids_od(IDs, ages, sex, ave_fecun, overdispersion)
}
\arguments{
\item{IDs}{vector of IDs}

\item{ages}{vector of ages}

\item{sex}{vector of individuals sex}

\item{ave_fecun}{the average fecundity at age (matrix one column for each sex)}

\item{overdispersion}{overdispersion parameter (one for each sex)}
}
\description{
Generate the list of kids born this year
}
