% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simhelpers.R
\name{get_growth}
\alias{get_growth}
\title{Find deterministic growth rate of Leslie Matrix}
\usage{
get_growth(Les)
}
\arguments{
\item{Les}{the Leslie matrix to find the growth rate of}
}
\description{
Find deterministic growth rate of Leslie Matrix
}
