% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simhelpers.R
\name{get_pars_with_target}
\alias{get_pars_with_target}
\title{Find missing parameter for target growth rate}
\usage{
get_pars_with_target(target)
}
\arguments{
\item{target}{the target deterministic growth rate to hit}
}
\description{
Find missing parameter for target growth rate
}
