% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simhelpers.R
\name{make_les}
\alias{make_les}
\title{Make Leslie Matrix}
\usage{
make_les(fecun, surv)
}
\arguments{
\item{fecun}{the fecundity vector}

\item{surv}{the survival vector}
}
\description{
Creates a simple Leslie matrix from fecundity and survival vectors
}
