% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simhelpers.R
\name{n_at_age}
\alias{n_at_age}
\title{Get numbers at age matrix from population}
\usage{
n_at_age(pop)
}
\arguments{
\item{pop}{the population to get numbers at age for}
}
\description{
Get numbers at age matrix from population
}
