% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsim.R
\name{rough_sample_size}
\alias{rough_sample_size}
\title{Generate rough sample size to target}
\usage{
rough_sample_size(
  tN,
  fun = function(y) {
     10 * sqrt(y)
 }
)
}
\arguments{
\item{tN}{matrix of true numbers at age}

\item{fun}{the function to use with}
}
\description{
Generate rough sample size to target
}
