% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsim.R
\name{samp_markers}
\alias{samp_markers}
\title{Sample markers from the population for breeding}
\usage{
samp_markers(smtlist, n)
}
\arguments{
\item{smtlist}{the list of markers used in the simulation}

\item{n}{the number to get back}
}
\description{
Sample markers from the population for breeding
}
