% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsim.R
\name{samp_non_fatally}
\alias{samp_non_fatally}
\title{Sample yearly non-fatally}
\usage{
samp_non_fatally(sim, sampquota, samp_years)
}
\arguments{
\item{sampquota}{the number at each age class to sample}

\item{samp_years}{the number of years to sample}

\item{population}{the population to sample}
}
\description{
This samples yearly non-fatally. It's meant to be run AFTER
the simulation is complete as it randomly samples based on alive
individuals in each of the years.
}
