% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsim.R
\name{sample_age_by_year}
\alias{sample_age_by_year}
\title{Sample by age by year non-lethally}
\usage{
sample_age_by_year(population, samp_size, sample_years, max_age)
}
\arguments{
\item{population}{data.table of population from simulation}

\item{samp_size}{the matrix of the sample size each year at age}

\item{sample_years}{vector of years from the sim to sample from39}

\item{max_age}{the max age of individuals in the population}
}
\description{
Sample by age by year non-lethally
}
