% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varCKMR.R
\name{simulate_markers}
\alias{simulate_markers}
\title{Simulate from markers dataframe}
\usage{
simulate_markers(markers, n)
}
\arguments{
\item{markers}{dataframe with the marker information geno,locus then freqs}

\item{n}{The number to simulate.}
}
\description{
Simulate from markers dataframe
}
