% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simhelpers.R
\name{surv_inv_fun}
\alias{surv_inv_fun}
\title{Keeps survival within a specified range}
\usage{
surv_inv_fun(y, a, b)
}
\arguments{
\item{y}{a value}

\item{a}{the lower part of the range}

\item{b}{the upper part of the range}
}
\description{
Keeps survival within a specified range
}
