% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsim.R
\name{yearly_deaths}
\alias{yearly_deaths}
\title{Death function}
\usage{
yearly_deaths(population, year_now)
}
\arguments{
\item{population}{the population to have individuals die in}

\item{surv_fun}{the survival function giving the probability of an individual making it to next year}
}
\description{
Given a population and survival function choose who dies
}
